Move tests to their own namespace and use PSR-4 autoloader
