+++ phpunit.xml.dist
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<phpunit bootstrap="tests/bootstrap.php" colors="true" strict="true" verbose="true">
+<phpunit bootstrap="vendor/autoload.php" colors="true" strict="true" verbose="true">
     <testsuites>
         <testsuite name="NinjaMutex Test Suite">
-            <directory>tests/NinjaMutex/</directory>
+            <directory>tests</directory>
         </testsuite>
     </testsuites>
