+++ tests/AbstractTest.php
@@ -7,20 +7,20 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace NinjaMutex;
+namespace NinjaMutex\Tests;
 
 use NinjaMutex\Lock\FlockLock;
-use NinjaMutex\Lock\MemcacheLock;
 use NinjaMutex\Lock\MemcachedLock;
+use NinjaMutex\Lock\MemcacheLock;
 use NinjaMutex\Lock\MySqlLock;
-use NinjaMutex\Lock\Fabric\MemcacheLockFabric;
-use NinjaMutex\Lock\Fabric\MemcachedLockFabric;
-use NinjaMutex\Mock\MockMemcache;
-use NinjaMutex\Mock\MockMemcached;
-use NinjaMutex\Mock\MockPredisClient;
 use NinjaMutex\Lock\PredisRedisLock;
-use Predis;
+use NinjaMutex\Tests\Lock\Fabric\MemcachedLockFabric;
+use NinjaMutex\Tests\Lock\Fabric\MemcacheLockFabric;
+use NinjaMutex\Tests\Mock\MockMemcache;
+use NinjaMutex\Tests\Mock\MockMemcached;
+use NinjaMutex\Tests\Mock\MockPredisClient;
 use org\bovigo\vfs;
+use Predis;
 
 abstract class AbstractTest extends \PHPUnit_Framework_TestCase
 {
@@ -138,7 +138,7 @@ abstract class AbstractTest extends \PHPUnit_Framework_TestCase
      */
     protected function provideMysqlMockLock()
     {
-        return array(new MySqlLock('', '', '', 'NinjaMutex\Mock\MockPDO'));
+        return array(new MySqlLock('', '', '', 'NinjaMutex\Tests\Mock\MockPDO'));
     }
 
     /**
