+++ tests/Lock/Fabric/LockFabricWithExpirationInterface.php
@@ -7,7 +7,8 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace NinjaMutex\Lock\Fabric;
+namespace NinjaMutex\Tests\Lock\Fabric;
+
 use NinjaMutex\Lock\LockInterface;
 use NinjaMutex\Lock\LockExpirationInterface;
