+++ tests/Mock/MockPDOStatement.php
@@ -7,7 +7,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace NinjaMutex\Mock;
+namespace NinjaMutex\Tests\Mock;
 
 use PDO;
 use PDOStatement;
