+++ tests/MutexLocksTest.php
@@ -7,10 +7,11 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace NinjaMutex;
+namespace NinjaMutex\Tests;
 
 use NinjaMutex\Lock\LockAbstract;
 use NinjaMutex\Lock\LockInterface;
+use NinjaMutex\Mutex;
 
 /**
  * Tests for Mutex's Locks
