+++ tests/MutexTest.php
@@ -7,9 +7,11 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace NinjaMutex;
+namespace NinjaMutex\Tests;
 
-use NinjaMutex\Mock\MockLock;
+use NinjaMutex\Mutex;
+use NinjaMutex\Tests\Mock\MockLock;
+use NinjaMutex\UnrecoverableMutexException;
 
 /**
  * Tests for Mutex
