--- tests/bootstrap.php
@@ -1,12 +1,0 @@
-<?php
-/**
- * This file is part of ninja-mutex.
- *
- * (C) Kamil Dziedzic <arvenil@klecza.pl>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-$loader = require_once __DIR__ . "/../vendor/autoload.php";
-$loader->add('NinjaMutex\\', __DIR__);
